requirements.txt
fastapi
uvicorn[standard]
python-multipart
PyMuPDF
fpdf2
requests
python-dotenv